Required

The API requires you to create a subtype of AbstractPlayer with extensions of the methods below.

Types

CantStop.AbstractPlayer — Type

AbstractPlayer

An abstract type for a player. Subtypes of AbstractPlayer must have the fields described below.

Fields

id::Symbol: player id

pieces::Vector{Symbol}: a vector of 12 pieces where each piece has the value id.

piece_reserve::Vector{}: an optional vector for keeping track of pieces which will replace runners unless a bust occurs

In addition, for a subtype MyPlayer <: AbstractPlayer, the API requires the following constructor to ensure the correct number of pieces are provided.

Constructor

MyPlayer(;id, pieces=fill(id, 12)) = MyPlayer(id, pieces)

Methods

CantStop.select_runners — Function

select_runners(::Type{<:AbstractGame}, player::AbstractPlayer, board, outcome, num_roll)

During the selection phase, select runners based on the outcome of a dice roll.

Arguments

::Type{<:AbstractGame}: an abstract game type

player::AbstractPlayer: an subtype of a abstract player

outcome: outcome of dice roll

board::Dict{Int,T}: a dictionary representing columns 2-12. Each row in a column is a vector of symbols which contain the player ids

num_roll: a count of dice rolls for a given round. Max is 2.

Returns

c_idx::Vector{Int}: a vector of column indices

r_idx::Vector{Int}: a vector of row indices

CantStop.take_chance — Function

take_chance(::Type{<:AbstractGame}, player::AbstractPlayer, board)

During the decision phase, decide whether to take a chance to advance the runners, or set your pieces in the current location of the runners.

Arguments

::Type{<:AbstractGame}: an abstract game type

player::AbstractPlayer: an subtype of a abstract player

board::Dict{Int,T}: a dictionary representing columns 2-12. Each row in a column is a vector of symbols which contain the player ids

Returns

decision::Bool: true if take chance, false otherwise

CantStop.select_positions — Function

select_positions(::Type{<:AbstractGame}, player::AbstractPlayer, board, outcome)

During the decision phase and after deciding to take chance, select positions based on the outcome of rolling dice. Two column indices are determined by summing two pairs of dice.

Arguments

::Type{<:AbstractGame}: an abstract game type

player::AbstractPlayer: an subtype of a abstract player

outcome: outcome of dice roll

board::Dict{Int,T}: a dictionary representing columns 2-12. Each row in a column is a vector of symbols which contain the player ids

Returns

c_idx::Vector{Int}: a vector of column indices

r_idx::Vector{Int}: a vector of row indices

CantStop.postbust_cleanup! — Function

postbust_cleanup!(::Type{<:AbstractGame}, player::AbstractPlayer)

Performs cleanup and book keeping after a "bust" (i.e., a roll that does not provide a valid move).

Arguments

::Type{<:AbstractGame}: an abstract game type

player::AbstractPlayer: an subtype of a abstract player

Returns

nothing

CantStop.poststop_cleanup! — Function

poststop_cleanup(::Type{<:AbstractGame}, player::AbstractPlayer)

Performs cleanup and book keeping after deciding to stop rolling during the decision phase.

Arguments

::Type{<:AbstractGame}: an abstract game type

player::AbstractPlayer: an subtype of a abstract player

Returns

nothing

Optional Types

It is possible to create variations of the game by creating a new subtype of AbstractGame and defining new internal methods as needed.

CantStop.AbstractGame — Type

AbstractGame

An abstract game type for Can't Stop.

The following fields are required in order to work with default methods:

Fields

dice::Dice: an object resepresenting four dice

board::Dict{Int,T}: a dictionary representing columns 2-12. Each column is a vector of symbol vectors which contain the player ids

c_idx::Vector{Int}: column indices of starting position of active piece

r_idx::Vector{Int}: row indices of starting position of active pieace

pieces::Dict{Symbol,Vector{Symbol}}: inactive pieces for each player: player_id -> pieces

piece_reserve::Vector{Symbol}: holds pieces for runners started at the beginning of the column

Available Types

CantStop.Game — Type

Game{T} <: AbstractGame

Fields

dice::Dice: an object resepresenting four dice

board::Dict{Int,T}: a dictionary representing columns 2-12. Each row in a column is a vector of symbols which contain the player ids

c_idx::Vector{Int}: column indices of starting position of active piece

r_idx::Vector{Int}: row indices of starting position of active pieace

pieces::Dict{Symbol,Vector{Symbol}}: inactive pieces for each player: player_id -> pieces

Constructor

Game(;dice=Dice(), board=make_board())

Utilities

The following methods might be helpful for defining the required methods above.

CantStop.get_active_locations — Function

get_active_locations(board, player_id)

Returns the location of active pieces, which includes runners.

Arguments

board::Dict{Int,T}: a dictionary representing columns 2-12. Each row in a column is a vector of symbols which contain the player ids

player_id::Symbol: id of player

Returns

c_idx: column indices of active positions

r_idx: row indices of active positions

CantStop.get_runner_locations — Function

get_runner_locations(board)

Returns runner locations.

Arguments

board::Dict{Int,T}: a dictionary representing columns 2-12. Each row in a column is a vector of symbols which contain the player ids

Returns

c_idx: column indices

r_idx: row indices

Missing docstring.

Missing docstring for is_valid_runner. Check Documenter's build log for details.

CantStop.list_sums — Function

list_sums(outcome)

Lists all unique sum of combinations of the outcome of rolling dice

Arguments

outcome: the results of rolling the dice

CantStop.reserve_piece! — Function

reserve_piece!(player::AbstractPlayer)

Tranfer piece to reserve for bookkeeping.

Arguments

player::AbstractPlayer: an object which is a subtype of AbstractPlayer

CantStop.simulate — Function

simulate(game::AbstractGame, players)

Simulate CantStop until a player has won.

Arguments

game::AbstractGame: an abstract game object

players: a vector of players where elements are a subtype of a abstract player