CantStop.clear_runners! — Method

clear_runners!(game)

Clear runners from board following a bust.

Arguments

game::AbstractGame: an abstract game object

CantStop.decision_phase! — Method

decision_phase!(game::AbstractGame, player::AbstractPlayer)

Implements the decision phase in which the player decides to roll the dice for the possibility of moving the runners. The two methods named decide! are called during this phase.

Arguments

game::AbstractGame: an abstract game object

player::AbstractPlayer: an subtype of a abstract player

CantStop.is_combination — Method

is_combination(outcome, columns)

Tests whether columns are a possible some of dice outcome.

Arguments

outcome: a vector of dice outcomes

c_idx: a vector of sums based on pairs of dice outcomes

Keywords

fun=all: a function for testing the combinations

CantStop.move! — Method

move!(game::AbstractGame, c_idx, r_idx)

Move runner to location determined by column and row index

Arguments

game::AbstractGame: an abstract game object

c_idx: column index

r_idx: row index

CantStop.play_round! — Method

play_round!(game::AbstractGame, player::AbstractPlayer)

Play one round with a specified player.

Arguments

game::AbstractGame: an abstract game object

player::AbstractPlayer: an subtype of a abstract player

CantStop.remove_pieces! — Method

remove_pieces!(game::AbstractGame, player_id)

Arguments

game::AbstractGame: an abstract game object

player::AbstractPlayer: an subtype of a abstract player

CantStop.replace_runners! — Method

replace_runners!(game::AbstractGame, player_id)

Replaces runners with player_id.

Arguments

game::AbstractGame: an abstract game object

player::AbstractPlayer: an subtype of a abstract player

CantStop.runner_selection_phase! — Method

runner_selection_phase!(game::AbstractGame, player::AbstractPlayer)

Implements the runner selection phase in which the player performs two dice rolls to select runners. The function select_runners! is called during this phase.

Arguments

game::AbstractGame: an abstract game object

player::AbstractPlayer: an subtype of a abstract player

CantStop.setup! — Method

setup!(player::AbstractPlayer, ids)

Perform initial setup after cards are delt, but before the game begins.

Arguments

player: a player object

ids: all player ids