Required

The API requires you to create a subtype of AbstractPlayer with extensions of the methods below.

Types

CantStop.AbstractPlayer — Type

AbstractPlayer

An abstract type for a player. Subtypes of AbstractPlayer must have the fields described below.

Fields

id::Symbol: player id

pieces::Vector{Symbol}: a vector of 11 pieces where each piece has the value id.

piece_reserve::Vector{Symbol}: an optional vector for keeping track of pieces which will replace runners unless a bust occurs

In addition, for a subtype MyPlayer <: AbstractPlayer, the API requires the following constructor to ensure the correct number of pieces are provided.

Constructor

MyPlayer(;id, pieces=fill(id, 11)) = MyPlayer(id, pieces)

Methods

The API requires one to define the following methods for each custom player type.

CantStop.roll_again — Function

roll_again(Game::AbstractGame, player::AbstractPlayer)

During the decision phase, decide whether to take a chance to advance the runners, or set your pieces in the current location of the runners.

Arguments

game::AbstractGame: an abstract game object for Can't Stop. The object is a copy, which means it is safe to modify.

player::AbstractPlayer: an subtype of a abstract player

Returns

decision::Bool: true if take chance, false otherwise

CantStop.select_runners — Function

select_runners(Game::AbstractGame, player::AbstractPlayer, options)

During the selection phase, select runners based on the outcome of a dice roll.

Arguments

game::AbstractGame: an abstract game object for Can't Stop. The object is a copy, which means it is safe to modify.

player::AbstractPlayer: an subtype of a abstract player

options: a vector of columns that can be selected based on outcome of rolling dice

Returns

c_idx::Vector{Int}: a vector of selected column indices for moving runner

CantStop.postbust_cleanup! — Function

postbust_cleanup!(Game::AbstractGame, player::AbstractPlayer)

Performs cleanup and book keeping after a "bust" (i.e., a roll that does not provide a valid move).

Arguments

game::AbstractGame: an abstract game object for Can't Stop. The object is a copy, which means it is safe to modify.

player::AbstractPlayer: an subtype of a abstract player

Returns

nothing

CantStop.poststop_cleanup! — Function

poststop_cleanup!(Game::AbstractGame, player::AbstractPlayer)

Performs cleanup and book keeping after deciding to stop rolling during the decision phase.

Arguments

game::AbstractGame: an abstract game object for Can't Stop. The object is a copy, which means it is safe to modify.

player::AbstractPlayer: an subtype of a abstract player

Returns

nothing

Optional Types

It is possible to create variations of the game by creating a new subtype of AbstractGame and defining new internal methods as needed.

CantStop.AbstractGame — Type

AbstractGame

An abstract game type for Can't Stop.

The following fields are required in order to work with default methods:

Fields

dice::Dice: an object resepresenting four dice

pieces::Dict{Symbol,Vector{Symbol}}: inactive pieces for each player: player_id -> pieces

columns_won::Vector{Int}: a vector of indices for columns won

max_rows::Dict{Int,Int}: maximum number of rows for each column

players_won::Dict{Int,Symbol}: indicates which player won a given column. A value of :_ indicates no player has won the column

Available Types

CantStop.Game — Type

Game{P<:AbstractPiece} <: AbstractGame

The default game object for CantStop.

Fields

dice::Dice: an object resepresenting four dice

pieces::Dict{Symbol,Vector{Symbol}}: inactive pieces for each player: player_id -> pieces

columns_won::Vector{Int}: a vector of indices for columns won

max_rows::Dict{Int,Int}: maximum number of rows for each column

players_won::Dict{Int,Symbol}: indicates which player won a given column. A value of :_ indicates no player has won the column

Constructor

Game(;dice=Dice(), piece_type=Piece)

Utilities

The following methods might be helpful for defining the required methods above.

CantStop.simulate! — Function

simulate!(game::AbstractGame, players)

Simulate CantStop until a player has won.

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.get_winner — Function

get_winner(game::AbstractGame)

Returns a vector of winners. If there is a tie, all winners are included in the vector.

Arguments

game::AbstractGame: an abstract game object for Can't Stop