CantStop.Dice — Type

Dice

An object for rolling dice.

Fields

n::Int: number of rolls

sides::Int: number of sides per die

Constructors

Dice(4, 6)
Dice(;n=4, sides=6)

CantStop.decision_phase! — Method

decision_phase!(game::AbstractGame, player::AbstractPlayer)

Implements the decision phase in which the player decides to roll the dice for the possibility of moving the runners. The two methods named decide! are called during this phase.

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.handle_bust! — Method

set_runners_false!(game::AbstractGame, player)

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.handle_stop! — Method

set_runners_false!(game::AbstractGame, player)

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.initialize_pieces! — Method

set_runners_false!(game::AbstractGame, player)

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.is_bust — Method

is_bust(game::AbstractGame, options)

Checks whether a dice roll is a bust i.e., does not allow a valid move.

Arguments

game::AbstractGame: an abstract game object for Can't Stop

options: a vector of column indices corresponding to possible choices

CantStop.is_playing — Method

is_playing(game::AbstractGame)

Arguments

game::AbstractGame: an abstract game object for Can't Stop

CantStop.list_options — Method

list_sums(outcome)

Lists all unique sum of combinations of the outcome of rolling dice

Arguments

game::AbstractGame: an abstract game object for Can't Stop

outcome: the results of rolling the dice

CantStop.list_sums — Method

list_sums(outcome)

Lists all unique sum of combinations of the outcome of rolling dice

Arguments

outcome: the results of rolling the dice

CantStop.move! — Method

move!(game::AbstractGame, id, r_idx)

Move runner to location determined by column and row index

Arguments

game::AbstractGame: an abstract game object

id: the id of the player

c_idx: column index of position

CantStop.play_round! — Method

play_round!(game::AbstractGame, player::AbstractPlayer)

Play one round with a specified player. During each iteration of a round, the player makes two decision_phase

decide whether to roll the dice

decide which dice to pair and sum to select columns

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.return_to_start_position! — Method

return_to_start_position!(game::AbstractGame, player::AbstractPlayer)

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.set_runners_false! — Method

set_runners_false!(game::AbstractGame, player)

Arguments

game::AbstractGame: an abstract game object for Can't Stop

player::AbstractPlayer: an subtype of a abstract player

CantStop.set_status! — Method

set_status!(game::AbstractGame, id, c_idx)

Move runner to location determined by column and row index

Arguments

game::AbstractGame: an abstract game object

id: the id of the player

c_idx: column index of position